Fixes issue 467 Fixes issue 1306
Adds a new hook conflict option HOOKED_TRAMPOLINE_NO_HOOK.
On a hook conflict, it bails adding a native_exec_syscalls
hook and we just live without it.

Changes the default -native_exec_hook_conflict value for CI builds to
HOOKED_TRAMPOLINE_NO_HOOK, as it seems better to let the app run correctly
when the risk of DR messing up is low.  Native interception matters only
for AppInit, for the rare tool that goes native, or for any future attach
feature: all rare things.

Changes the default -handle_ntdll_modify for CI builds to
DR_MODIFY_ALLOW.  For CI, the goal is to run an arbitrary app and err on
the side of DR missing stuff while native rather than messing up the
app's behavior.

Fixes a bug in DR_MODIFY_ALLOW where it skips handling of overlap with
executable areas.

SVN-Revision: 2417