Fixes issue 587
Remove template parameters for PDB symbols in drsyms.

Removal is done under the assumption that return and function parameter
types are not present, which simplifies the operation.

For drsym_lookup_address(), drsym_enumerate_symbols_ex(), and
drsym_search_symbols_ex(): we were already copying the symbol name with
strncpy.  Now we walk the name and selectively copy, removing the template
parameters as we go.

For drsym_enumerate_symbols() and drsym_search_symbols() we are doing an
extra walk through the symbol name, an in-place copy (using dbghelp's own
buffer).

The empty "<>" are left in place, to match non-PDB behavior.

A caveat: querying dbghelp will still look at a template parameters.  Thus,
drsym_lookup_symbol() users must still specify them, and
drsym_search_symbols*() may find names that seemingly do not have the
target string in them as we removed it later.

Added a new flag DRSYM_DEMANGLE_PDB_TEMPLATES, to support disabling
this feature (in case it's buggy, or for those who want something in
between DRSYM_DEMANGLE and DRSYM_DEMANGLE_FULL, or for those who want
the behavior of the old drsyms).

Added DRSYM_DEMANGLE_PDB_TEMPLATES to symquery, as it seems we want as
much information as possible there.

Added tests to drsyms-test.

I measured the performance hit of this detemplatizing on every symbol
operation and it was negligible, as we were already copying the buffer for
nearly all operations.

Still TODO in i#1530: add a way to pass demangling flags to
drsym_search_symbols*() routines.

SVN-Revision: 2497