i#1359 port drsyms to Mac, part 1
Adds drsyms_macho.c, which walks Mach-O symbol tables, creates a sorted
array, and provides the drsyms obj interface for acccessing it.  Mach-O
symbols, like on PECOFF, have no stored size.

Adds 32-bit Mach-O versions of libelftoolchain's libdwarf.a and libelftc.a,
along with a patch and instructions on how to build them.  This allows
DWARF2 querying on MacOS.

SVN-Revision: 2516