Fixes issue 1367
Adds better support for non-contiguous modules by passing only the first
segment's endpoint to module_list_add(), where the range is required to be
disjoint from other modules.  We then update the stored endpoint to be the
max for a contiguous module, and leave it for non-contiguous.

Updates the docs to indicate that the end field in several module data
structures is just the first segment for non-contiguous modules.

SVN-Revision: 2528