Fixes issue 1371
On Mac, which has separate tables of system calls, not only is system call
number encoding used to distinguish the call type (like on Windows), but
also the gateway.  int 0x81 and int 0x82 use different tables from int
0x80.

This commit adds support for this:

1) Distinguish which type for non-ignorable syscalls via LINK_SPECIAL_EXIT
and new exit_reasons for 0x81 and 0x82.  This includes interp, mangling,
and dispatch support.

2) Generate do_int8{1,2}_syscall routines.

3) Pass more info to os_normalized_sysnum() and ignorable_system_call() to
identify the gateway.

SVN-Revision: 2551