Fixes issue 1382
Adds sub-xmm-register size support to the IR by using the opnd_t.size field
for REG_kind.  Simplifies the size use by setting it to 0 and supporting
that meaning the full register size.

Although we now expose sub-register sizes (i#1382), we do not require them
when encoding as we have no simple way to add auto-magic creation to the
INSTR_CREATE_ macros.  Plus, sub-register sizes never distinguish two
opcodes.

Updates many instructions to clarify sub-xmm-register usage, including
instruction classes that zero the top bits of the destination even though
they primarily operate on the low 4 or 8 bytes; non-destructive-source
classes that propagate the top bits of one of the sources; and other
miscellanous instructions.

Updates all of the corresponding test code to explicitly create partial
registers to pass the instr_same() test (one of the drawbacks of this
design is that the INSTR_CREATE_ macros don't auto-magically do this for
you).

SVN-Revision: 2578