i#58 MacOS: initial thread handling
Prelminary thread handling by intercepting at the user thread start passed
to SYS_bsdthread_create.  This is not ideal: it's not the first user code
(we miss the pthreads layer).  Plus, this won't intercept workqueue threads
at all.  What we want to do is intercept bsdthread_register, but we need
earlier injection for that.

This also tweaks the TLS assumption: we seem to need a new LDT entry for
new thread.  i#1405 covers investigating why the kernel isn't swapping our
entry as I originally thought it would.

SVN-Revision: 2616