i#1509: have drrun list available tools
There's no simple solution here so I went with an install-time-generated
text file (appended to separately by each tool, which requires coordination
with Dr. Memory as well) that names the tools we want to advertise (e.g.,
we want to not list the legacy name "bbcov").  At runtime drrun reads the
file and prints the info in its usage message.

Fixes issue 1509

SVN-Revision: 2768