i#1551 port to ARM: initial disassembly
Split disasm byte printing for ARM vs x86 styles and implement ARM and both
Thumb sizes.

Start implementing instr_is_* for ARM.

Add type_names[] for ARM.

Disassembly of immediates on ARM: decided to always show as positive or
negative, partly b/c of the lack of pointer-sized immeds on ARM.  Decided
not to extend this in any way to x86 to avoid breaking any existing uses of
the disassembler.

SVN-Revision: 2920