i#1551 port to ARM: register operand flags
Adds a new type, dr_register_flags_t, for representing additional
properties of a register operand.  The two flags now present are
DR_REGMOD_VALUE_NEGATED and DR_REGMOD_VALUE_SHIFTED which indicate changes
to a register's value when used in an instruction.  The negation is
necessary for addressing mode writeback and post-indexing, as we did not
split the opcodes.  The shift is a nice hint.

Renames opnd_t.seg to opnd_t.aux.
Adds opnd_create_reg_ex() and opnd_get_flags().
Adds ARM decoding of negated registers.

SVN-Revision: 2962