i#1551 port to ARM: implement instr creation backend for macros
Implements routines needed for the ARM INSTR_CREATE_ macros:
+ instr_create_0dst_4src()
+ instr_create_2dst_5src()
+ instr_create_Ndst_Msrc_vardst()
+ instr_create_Ndst_Msrc_varsrc()

Changes the macro script to add a parameter to the var-arg routines giving
the number of registers in the register list as there's no way to determine
that automatically.

Also fixes the macro script to add a missing opnd_create_reg() around
opnd_get_base().

SVN-Revision: 3014