i#1551 port to ARM: set dcontext ISA mode in IBL
As part of the ISA mode switch we must set the dcontext current mode in the
IBL (as there are too many paths out of the code cache to centralize in C
code).  We currently pay this cost every time we enter the IBL, even when
not switching modes, b/c we don't have 2 copies of the gencode.

Review-URL: https://codereview.appspot.com/199220043