i#1551 port to ARM: mangle app uses of the stolen register
Adds initial support for mangling the app's reads and writes of the stolen
register.  Adds simple local scratch register handling of register list
writes, register list reads, and normal instruction reads and writes.  Does
not consider all interactions with pc mangling and indirect branch
mangling: this is just a first pass.

Adds instr_remove_srcs() for use in this mangling, and exports it.

Review-URL: https://codereview.appspot.com/197280043