i#1551 port to ARM: fix IBL flag-clobbering bug and switch to all-Thumb gencode
For the ARM IBL, do not use OP_cmp as it clobbers flags.  Instead we use
OP_cbnz and require Thumb for our gencode.  We remove the -isa_mode_arm
runtime option, as we now fully support ARM<->Thumb transitions, which are
inexpensive enough that we do not need separate gencode versions.

Review-URL: https://codereview.appspot.com/200540043