i#1551 port to ARM: fix IT block reinstatement bug
Fix bug where we incorrectly insert an extra IT instr inside an IT block we
just added.

Adds a client-visible error in IT tracking on an OP_it instr with nothing
after it, but only if nothing is before it, so we don't complain if
instr_encode() is called on an instr not in an ilist.

Review-URL: https://codereview.appspot.com/207270043