i#1613 Finalize annotations: coordinating with client instrumentation
Since the client may remove instructions in an annotated
basic block, we have a new policy to determine whether to
instrument the annotation, so that the client can choose
to omit the annotation, and so that the instrumentation
does not accidentally get duplicated. In this commit, the
annotation instrumentation is "attached" to the preceding
app instruction, and will only appear where a translation
of that instruction appears (unless the annotation is at
the start of its basic block, in which case the
instrumentation is always applied).

Review-URL: https://codereview.appspot.com/207240043