i#1551 port to ARM: fix extra bits set in duplicate entries
Duplicates in tables where a distinguishing immed bit must be set for the
decoder to get there must not set that bit in their opcode field.  To
handle that we'd need logic in the chaining script to select the duplicate
with the fewest bits set, rather than just taking the first one.  An
example is OP_lsls which was listed as 0x01b00080 but which must be changed
to list 0x01b00000.

Review-URL: https://codereview.appspot.com/213720045