i#1551 port to ARM: refactor sigcontext code
For Linux on ARM, there is no pointer to the fp/simd state from the
sigcontext.  Thus we have to pass additional information everywhere.  We
refactor the code to pass in kernel_ucontext_t where available, and where
not we use a new struct sig_full_cxt_t which is exposed for use as the
suspended context for synch.

Review-URL: https://codereview.appspot.com/210300043