i#1551 port to ARM: support auto-restarting interrupted syscalls
On ARM, r0 holds EINTR and unlike xax on x86, it's also a param to the
syscall, so we have to save it before each syscall so we can restore in
order to restart.  Adds saving of r0 both for inlined syscalls and for
do_syscall.  Adds ignoring of the r0 save to the translation walk.

Implement instr_is_trace_cmp(), instr_is_inline_syscall_jmp(),
interrupted_inlined_syscall(), and adjust_syscall_for_restart() for ARM.

Enables the linux.eintr and linux.eintr-noinline tests, which now pass.

Review-URL: https://codereview.appspot.com/208600044