i#1541 library deps: remove libdl dependence
Makes -private_loader required on Linux in order to remove the libdl
dependence from libdynamorio.so.  We leave libdl for MacOS (no private
loader yet: i#1596) and STATIC_LIBRARY.

For -native_exec, we run the current tests with -mangle_app_seg to keep
things working, but long-term if we want to revive a hybrid mode we'll want
to add a separate build option for it where libdl is imported from.

Fixes #1541

Review-URL: https://codereview.appspot.com/211440043