i#1551 port to ARM: encode order bug fix
For T32 OP_ldr with MPCN12w opnd, we have to take the PC-base special
template first.  Otherwise, if the immed fits in 8 bits, we'll take a
template that has some other value in those 4 bits that the PC-base one
interprets as part of the immed.  My solution here is brute-force: raise
priority of all templates in PC-based tables.  It's simple, though, and
there shouldn't be any negative consequences.

Review-URL: https://codereview.appspot.com/215100043