i#1551 port to ARM: IT decode tracking in decode_from_copy()
The decode_from_copy() used in check_encode_decode_consistency() raises
fragility problems in decode IT block tracking: the state will get off if
our encoding doesn't match the app's in length, b/c we're advancing
according to app length while IT tracking will advance at our length.  This
happens for deliberate .w instrs in the app that have natural .n forms.

This is solved for this case via a workaround that adjusts the decode IT
state directly.  This breaks abstraction but it's all inside the decoder.
We should think about how to support this for clients in a cleaner way.

Adds a test to allasm_thumb which triggers this.

Review-URL: https://codereview.appspot.com/211530043