i#909 early execve: implements early injection across execve
Previously, even if the parent was invoked with -early_inject, any
subsequent execve was injected into using LD_PRELOAD.
This CL implements early injection across execve by substituting the
libdynamorio.so path for the app's target path (which requires adding new
filepath buffers).  We also replace its argv[0] if it's /proc/self/exe.

It turns out that our unsetenv during init shifts the env entries and
messes up ld.so's access to auxv.  We have to change those calls to a new
routine disable_env() which leaves the entries in place but changes their
names.

The linux.readlink test functions as a regression test for this.

Fixes #909

Review-URL: https://codereview.appspot.com/213290043