i#1522 AppInit on Win8: enable AppInit for WOW64
Changes drpreinject to not do its hacky self-unload on Win8+, to avoid
crashes and hangs at exit (even on un-configured processes) when
AppInit is enabled.

Combined with i#1035, this is enough to get -syswide_on working for
Win8+ WOW64, so this CL enables that in drconfig.  However, x64
AppInit seems to still have some issues, so further i#1522 work is
neeeded.

Review-URL: https://codereview.appspot.com/218960043