i#1632 intercept hook truncation: translate pc occluded by hook
When execution enters an intercept, the client-visible translation
for all intercept instructions is set to the original app pc,
which is occupied by the hook. If the client truncates this BB,
this adjusted translation may point into the middle of the hook,
resulting in decoding errors. This commit augments the mapping
intercept pc <-> app pc, adding a flag to indicate whether any
app instructions are occluded by the intercept hook. The map is
used to adjust the source pc and fall-through pc during
build_bb_ilist() (as necessary). This commit also extends the
client.truncate test to cover truncation at 1 or 2 instructions,
which will exercise this scenario, and adds an execution of the
truncate client on the win32.threadchurn app to exercise
additional cases of hook-occluded app instructions. This fixes
i#1610 which requires intercept truncation to work.

Fixes #1610

Review-URL: https://codereview.appspot.com/220080043