i#1589 .hash crash: handle DR loaded by system ld.so
Fixes a third issue: for libdynamorio.so itself, if it was loaded by the
system ld.so, its .dynamic *will* be relocated.

Adds a new test to the suite "linux.fib-conflict" which is built to always
conflict with libdynamorio.so's preferred base.  This was accomplished by
refactoring core/CMake_ldscript.cmake into make/ldscript.cmake and
generalizing it a bit.  This will be useful for i#1227 in the future.

Fixes #1589

Review-URL: https://codereview.appspot.com/217440043