i#1613 finalize annotations: minor corrections
A recent change to build_bb_ilist() assumed that all
annotations have a return placeholder, which is not
the case for void annotations. This commit also
clarifies the signature for the "is running on DR"
annotation, and sets the return value to zero for
non-void annotations when no handler is registered.

Review-URL: https://codereview.appspot.com/221240043