i#1004 early brk: emulate the brk
Adds a new option -emulate_brk, which is on by default for -early_inject,
but off by default otherwise.
When on, we emulate the app's brk via an mmap placed after the application
executable (modulo -separate_private_bss).  This emulation extends to
encompass dr_raw_brk() as well.

The option -no_emulate_brk is communicated to early injection via a new env
var set by drinjectlib, as we have not yet parsed the option string at that
point.

Fixes #1004

Review-URL: https://codereview.appspot.com/223830044