i#884 module event too early: move module load event to first execution
Moves the module load event to the first execution of code from the module,
to solve problems with the current load event being prior to relocations
and making it difficult for the client to properly examine and act on the
module.

Leaves the signature of the module load event unchanged, with the bool
param remaining for compatibility.

Adjusts when check_new_page_start() is called to give clients adding to the
null_instrument_list a chance to add a module on its first execution.
Bails on guaranteeing the same thing in the presence of elision for a bb
that crosses module boundaries.

Adds a check for executable_vm_area_executed_from() to
dr_delay_flush_region() to avoid flushes called from the module load event
from now flushing code when with the old at-map event they would not have.

Adds serialization of bb creation to thread-private configurations when a
module load event is present, and disallows changing module events during
bb building.

Fixes #884

Review-URL: https://codereview.appspot.com/218480043