i#1663 firefox hook conflict: remove LdrLoadDll hooks
Moves the block_mod_load_list implementation from LdrLoadDll to
presys_MapViewOfSection.  We need this feature to stop double
injection when using both AppInit and late injection.  This then
removes the need for the LdrLoadDll hook in a default build.

Adds a new option -hook_ldr_dll_routines, which is off by default,
controlling whether we hook LdrLoadDll and LdrUnloadDll.  We will also
hook if gbop is enabled or on Win2K.

Removing these hooks should improve interoperability.

Tested on custom AppInit plus follow-children injection which is not
easy to add to our suite unfortunately due to the admin privileges
required on Win7+.

Fixes #1663

Review-URL: https://codereview.appspot.com/228730043