i#1551 port to ARM: add -steal_reg_at_reset diagnostic feature
Adds a new option -steal_reg_at_reset which swaps which register we steal
mid-run.  It has to reset all caches, so it must be used in concert with
something like -reset_at_fragment_count.  At the reset, we re-create the
gencode and update the register values in the suspended threads.

Unfortunately there are still some bugs somewhere, likely in the core reset
handling itself: probably in ARM translation.  On larger applications we
see crashes with this feature about half the time, but working the other
half has been enough to make it very useful and worth committing.

Review-URL: https://codereview.appspot.com/228800043