i#1665 pause w/o stdin: adds -pause_via_loop option
-msgbox_mask currently relies on reading stdin, which fails on applications
that also read stdin.  We add a new option -pause_via_loop which waits in
an infinite loop (of thread yields, to avoid taking CPU) and allows
attaching to such applications.

This -pause_via_loop supersedes the old -pause_on_error which gave too
little control as it was tied to the dumpcore options instead of the msgbox
options.

Fixes #1665

Review-URL: https://codereview.appspot.com/228040043