i#1676 early inject under gdb: add warnings and documentation
Adds detection of the early injection ASLR bug in privload_early_inject()
by checking whether libdynamorio.so is at the telltale 0x5... address.  If
so, we cannot easily display a warning message as we cannot reference any
global variables, so we use a char array on the stack to print a quick
message and then exit, rather than crashing.

Adds detection of running inside gdb to drrun and drinject.  Just a warning
is printed as we do not know whether the user has already run the proper
gdb command.

Adds documentation of the command to run in order to launch drrun from gdb.

Fixes #1676

Review-URL: https://codereview.appspot.com/226950043