i#1683 arm disasm: register lists
Adds a new operand flag DR_OPND_IN_LIST and has the decoder set it as well
as the vararg instr creation routines set it.  This makes disassembling
with braces around a register list much easier.  For now we do not collapse
adjacent registers into the "rN-rM" form, as not all decoders do that and
thus it seems unnecessary.

Refactors instr_disassemble_opnds_noimplicit() to avoid encoding and avoid
crashing on register lists trying to get the table template type as we do
not (yet) need that for ARM disassembly.

Review-URL: https://codereview.appspot.com/231020043