i#1673 port ext to ARM: add replace_native support to ARM
Adds replace_native support for ARM.  Fixes issues with calls involving a
link register rather than a return address pushed on the stack and other
miscellaneous porting steps.

Enables the drwrap test in the test suite and ports the assembly within
the test to ARM.

Related routine dr_call_on_clean_stack() is not yet ported over: it is left
for future work.

Review-URL: https://codereview.appspot.com/233260043