i#1723 stack problems: fix off-by-one query on stack top
Fixes an off-by-one query where we ask about the top of the app stack,
which can in fact be a DR address if the next page is DR's.  This can
lead to a zeroed out TEB.StackBase and subsequent Dr. Memory asserts
or worse.  We need to query one byte below the top.

Review-URL: https://codereview.appspot.com/237980043