i#1703 cache simulator: initial tracer client
Finishes the initial end-to-end named pipe framework for drcachesim by
adding the tracer client.  This initial client is based on
memtrace_simple.c except it writes memref data to a pipe, and uses a shared
memref data layout.  The pipe is a single global pipe, per our simplified
design.

Officially adds drcachesim to the build process for Linux (the Windows
named pipe code is NYI).

Review-URL: https://codereview.appspot.com/246080043