i#1705 option parsing: easy-to-use C++ options via droption extension
Adds a new Extension called "droption" which provides easy-to-use C++
options for both clients and standalone applications.  The user simply
declares instances of droption_t, calls dr_parse_options(), and that's it.
Static initializers and partial specialization take care of the rest,
including default value support, minimum and maximum range values, and
various ways to negate boolean options.  Error handling is passed back to
the user with a string describing the error.

Modifies the option_parse test client to be C++ and to test droption.
Adds documentation and release notes.

I still want to add a usage message feature so this does not yet close the
issue.

Review-URL: https://codereview.appspot.com/248150043