i#1705 option parser: change scope to bitfields; add get_name()
Changes the option scopes from an enum to bitfields, to make it easier to
check for combinations of client and frontend.  Also fixes a bug (due to
the old enum approach being awkward) where the passed-in scope being the
combination failed to match all requested scopes.

Adds get_name() so users can manipulate option keys.

Also fixes a bug in the boolean option parsing code.

Review-URL: https://codereview.appspot.com/242640043