i#1703 cache simulator: parametrize the cache configuration
Adds new runtime options for specifying the cache configuration, which for
now is limited to a simple 2-level hierarchy with a single line size.  The
parameters look like this:

-cores 4 -line_size 64 -L1I_size 32K -L1I_assoc 8 -L1D_size 23K -L1D_assoc 8
-LL_size 8M -LL_assoc 16

More complex cache layouts are covered by future work in i#1715 where we
would take in a configuration file.

Threads are assigned to cores using modulus for now: this will be changed
shortly.

Review-URL: https://codereview.appspot.com/254720043