i#1703 cache simulator: load-balanced thread-to-core mapping
Adds a fairer thread-to-core mapping.  We still have a static scheduling,
but we place each new thread on the core with the fewest assigned threads.

Adds a new trace entry type to indicate a thread exit, which unlike the
other meta-entries is passed through to the simulator by the IPC reader.

Adds statistics on the number of threads that executed on each core, and
adjusts the stats output for better per-core grouping.

Review-URL: https://codereview.appspot.com/248480043