i#1703 cache simulator: add initial tests
Adds two end-to-end black-box tests of drcachesim.
Both run existing tests and check for output being in the ballpark of what
I see on my machine: hello,world and client.annotation-concurrency (a nice
multi-threaded test).  These may end up being flaky and we may want to
remove them once we have targeted tests but they should be a solid initial
pair of sanity checks for drcachesim.

Makes use of the new test refactoring to run the tests via the -t option to
drrun.

Review-URL: https://codereview.appspot.com/249570043