i#1703 cache simulator: automated option html documentation
Adds initial drcachesim documentation with a focus on the runtime options.

Adds prefix and suffix parameters to droption's usage_long() and adds a
helper program that, when run, generates html option usage automatically.
The helper is built and run on Windows as well, although drcachesim itself
is not supported there (i#1727).

Adds CMake support with a custom target and dependencies and doxygen source
list updates throughout the documentation generation chain in order to run
the helper program and pass the build-time-generated docs to doxygen for
the final DynamoRIO docs.

Review-URL: https://codereview.appspot.com/250280043